#17 Add basic GUI
The GUI is using PyQt5 and interacting with the simulation through the
FastAPI interface (localhost:8000). It includes buttons to start and
stop the simulation and a basic interactive map of Wrocław created
with folium. It runs separately from the simulation server, but for
convenience you can use `poetry run app` to start both of them at the
same time.